Fixed bug in integer size for large ematricies

+++ dimreduce.c
@@ -167,18 +167,18 @@ int do_dimreduce(int argc, char *argv[], int mpi_id, int mpi_num_procs) {
   // Calculate the total number of comparisions and how many will be
   // performed by this process. We subtract 1 from the first params.rows
   // because we do not calculate the diagnoal.
-  int num_rows = params.rows - 1;
-  long int total_comps  = num_rows * (num_rows + 1) / 2;
-  long int comps_per_process = total_comps / mpi_num_procs;
-  long int comp_start = mpi_id * comps_per_process;
-  long int comp_stop = mpi_id * comps_per_process + comps_per_process;
+  long long int num_rows = params.rows - 1;
+  long long int total_comps  = num_rows * (num_rows + 1) / 2;
+  long long int comps_per_process = total_comps / mpi_num_procs;
+  long long int comp_start = mpi_id * comps_per_process;
+  long long int comp_stop = mpi_id * comps_per_process + comps_per_process;
 
   // If this is the last process and there are some remainder comparisions
   // then we need to add them to the stop
   if (mpi_id + 1 == mpi_num_procs) {
     comp_stop = total_comps;
   }
-  printf("%d. Performing %ld comparisions\n", mpi_id + 1, comp_stop - comp_start);
+  printf("%d. Performing %lld comparisions\n", mpi_id + 1, comp_stop - comp_start);
   fflush(stdout);
 
   // Perform the pair-wise royston test and clustering
